removed extra code with little refactoring
